This checkin, and the next few, should really have been in a branch, because I might want to back them out later.  But anyway:
Beginnings of a new idea -- in predict-score, use a _fresh_ history
rather than dragging along the current history.  

It's faster: "make check" reports about 90 seconds versus 120 seconds
for the last revision.

Now I need to change the code that determines if cards are in
sequence: right now they're looking in the (now-empty) history, which
is silly; they should instead look in the enemy's hands.