Fix one subtle bug, but notice another.
The fixed bug: when considering if a sequence of cards was adjacent,
we were failing to take into account cards that the enemy had played
_to the current trick_.

The bug I just noticed:

    Trick 2:
    east: (h2 s3 d4 s4 s5 h6 s8 ct dj sq dk sk)
    -> ((h2) (s3) (d4) (s4 s5) (h6) (s8) (ct) (dj) (sq) (dk) (sk))

Note that the d4 is in between the s3 and the s4.  That's crazy.  The
right grouping is more like

    ((h2) (d4) (s3 s4 s5) (h6) (s8) (ct) (dj) (dk) (sq sk))

This bug causes us to have too many groups, and that'll _kill_
performance, since it potentially requires more recursive calls to
predict-score.