Working on the previously-mentioned bug:
play-loop is now more general.  Rather than taking a fixed number of
tricks to play, it instead takes a proc that tells it when to stop.
(Actually it already took a proc which told it what to do _when_ it
stopped; that proc now does both duties).

One disturbing observation: the first line of spew contains "Oh good,
n knows for certain that s6 yields 1".  That seems bogus; clearly the
s6 is a loser, not a winner.  On the other hand, n ultimately plays
the d2, which is a better choice (since s has the da)