beginnings of better testing for planet-emacsen.ss
At first I was stumped because it seemed to be "snarfing" atom data
once per minute, instead of once every ten seconds as I'd have
expected.  Then I realized that that was because I'd limited the size
of the async-channel; that channel would fill up, and "puts" to it
would block until the bot drained the items.